--- FinalProject/src/main/java/com/team12/main/t2/regProduct/RegProductMapper.java
@@ -1,14 +1,0 @@
-package com.team12.main.t2.regProduct;
-
-public interface RegProductMapper {
-
-
-	void regProduct(Product p);
-
-
-
-
-	
-
-
-}
