--- FinalProject/src/main/java/com/team12/main/t2/viewProductPage/ViewAllProductMapper.java
@@ -1,13 +1,0 @@
-package com.team12.main.t2.viewProductPage;
-
-import java.util.List;
-
-import com.team12.main.t2.regProduct.Product;
-
-public interface ViewAllProductMapper {
-	
-	public List<Product> getAllProduct();
-
-
-
-}
