--- FinalProject/src/main/resources/team2/YJ/regProductMapper.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.team12.main.t2.regProduct.RegProductMapper">
-
-	
-
-	<insert id="regProduct" parameterType="com.team12.main.t2.regProduct.Product">
-		INSERT into PRODUCT VALUES(Product_seq.NEXTVAL,#{pet_category},#{toy_category},#{productName},#{productPrice},#{productInfo},#{productStock},#{onExhibition},#{productThumbnail},#{productImg},#{productTag})
-	</insert>
-	
-
-</mapper>
