--- FinalProject/src/main/resources/team2/YJ/viewAllProductMapper.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.team12.main.t2.viewProductPage.ViewAllProductMapper">
-	
-	<select id="getAllProduct" resultType="com.team12.main.t2.regProduct.Product">
-		select * from PRODUCT
-	</select>
-
-
-</mapper>
