+++ FinalProject/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -50,8 +50,7 @@
 		<beans:property name="mapperLocations">
 			<beans:list>
 				<beans:value>classpath:testmapper.xml</beans:value>
-				<beans:value>classpath:team2/YJ/viewAllProductMapper.xml</beans:value>
-				<beans:value>classpath:team2/YJ/regProductMapper.xml</beans:value>
+				<beans:value>classpath:team2/YJ/productMapper.xml</beans:value>
 				<beans:value>classpath:team1joinMapper.xml</beans:value>
 				<beans:value>classpath:team2BoardMapper.xml</beans:value>
 				<beans:value>classpath:team2loginMapper.xml</beans:value>
