+++ FinalProject/src/main/webapp/WEB-INF/views/2Team/t2_index.jsp
@@ -46,7 +46,7 @@
 	<div id="indexMenu">
         <ul>
             <li><a href="hospitalGo">동물병원 찾기</a></li>
-           <li><a href="viewProductPage">상품 판매</a></li>
+           <li><a href="viewProductPage.go">상품 판매</a></li>
             <li class="t2-indexDropdown">
             	<a href="#" class="t2-indexDropbtn">게시판</a>
             	<ul class="t2-indexDropdown-content">
